# docs/swagger/api/swagger.yaml
basePath: /
definitions:
  account.Account:
    properties:
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      shard_key:
        type: string
      token:
        type: string
      uuid:
        type: string
    type: object
  error.Error:
    properties:
      error_message:
        type: string
      status:
        type: integer
    type: object
...
  github.com_game-core_gocrafter_api_presentation_request_account.RegisterAccount:
    properties:
      name:
        type: string
    type: object
...
  github.com_game-core_gocrafter_api_presentation_response_account.RegisterAccount:
    properties:
      account:
        $ref: '#/definitions/account.Account'
      status:
        type: integer
    type: object
...
host: localhost:8001
info:
  contact: {}
  description: This is a sample swagger server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: github.com/game-core/gocrafter
  version: "1.0"
paths:
...
  /account/register_account:
    post:
      consumes:
      - application/json
      parameters:
      - description: アカウント登録
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github.com_game-core_gocrafter_api_presentation_request_account.RegisterAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_game-core_gocrafter_api_presentation_response_account.RegisterAccount'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.Error'
      summary: アカウント登録
      tags:
      - Account
...
swagger: "2.0"